+++ src/main/java/fr/eseoye/eseoye/beans/PostComplete.java
@@ -8,12 +8,12 @@ import java.sql.Date;
 public class PostComplete extends Post {
     private String content;
 
-    public PostComplete(String id, String title, User author, int price, Date date, String content) {
+    public PostComplete(String id, String title, User author, float price, Date date, String content) {
         super(id, title, author,price, date);
         this.content = content;
     }
 
-    public PostComplete(String title, User author, int price, Date date, String content) {
+    public PostComplete(String title, User author, float price, Date date, String content) {
         super(title, author,price, date);
         this.content = content;
     }
